+++ CHANGES.md
@@ -37,6 +37,8 @@
 * Add basic support to read overviews
 * BREAKING: update geo-types to 0.7.0. geo-types Coordinate<T> now implement `Debug`
   * <https://github.com/georust/gdal/pull/146>
+* Deprecated `SpatialRef::get_axis_mapping_strategy` - migrate to
+  `SpatialRef::axis_mapping_strategy` instead.
 
 ## 0.7.1
 * fix docs.rs build for gdal-sys